Player Upgrades, Coin System, UI Manager, and Critical Chance Behavior
 Added buttons for attack, defense, and utility upgrades that the player can purchase with coins. Currently, only the attack upgrades are functioning.
Enemies now drop coins when defeated. The player is rewarded with the amount of coins dropped, which can be used for upgrades.
 Integrated a UI manager to assist with displaying important messages. It handles showing the confirmation message for upgrades, as well as a message when the player doesn’t have enough coins.
Implemented a critical chance mechanic for the player's bullets, allowing for a chance to deal extra damage when a critical hit occurs. (2.5 X damage)

Issues:
Not an issue, but note that the player can damage multiple enemies at once (similar to a shotgun, so it may be ok ). (commit 3 - dont think about it yet)
The player still shoots at dead enemies (commit 2).